# apps/api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Edutech Platform API",
        "description": "API documentation for Edutech Platform - Bimbel & Tryout Platform",
        "contact": {
            "name": "Edutech Team",
            "email": "api@edutech.example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/api/v1/exams/packages": {
            "get": {
                "tags": [
                    "CBT"
                ],
                "summary": "List exam packages",
                "description": "Retrieves a list of available exam packages.",
                "operationId": "6851d312e952afffa58edd9008bf72ee",
                "responses": {
                    "200": {
                        "description": "Packages retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Packages retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Tryout SNBT 2025"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Description here"
                                                    },
                                                    "duration": {
                                                        "type": "integer",
                                                        "example": 120
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "active"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/exams/packages/{id}": {
            "get": {
                "tags": [
                    "CBT"
                ],
                "summary": "Get exam package details",
                "description": "Retrieves details for a specific exam package.",
                "operationId": "c2fe64af1a6ab519b8e2cfd5e557b87b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Package ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Package details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Package details retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Tryout SNBT 2025"
                                                },
                                                "level": {
                                                    "type": "string",
                                                    "example": "sma"
                                                },
                                                "duration_minutes": {
                                                    "type": "integer",
                                                    "example": 120
                                                },
                                                "randomize": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "show_result_mode": {
                                                    "type": "string",
                                                    "example": "after"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Package not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/exams/{attempt}/submit": {
            "post": {
                "tags": [
                    "CBT"
                ],
                "summary": "Submit exam",
                "description": "Submits the exam, calculates the score, and finishes the session.",
                "operationId": "e6d922ed5527251b43e7a5d564d65f95",
                "parameters": [
                    {
                        "name": "attempt",
                        "in": "path",
                        "description": "Exam attempt ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exam submitted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Exam submitted successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "score": {
                                                    "type": "integer",
                                                    "example": 85
                                                },
                                                "submitted_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Exam already submitted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}